api/pyproject.toml
[tool.poetry]
package-mode = false
name = "data-sharing-barter-incentives-rest-api"
version = "0.1.0"
description = "Restful API and Database for INESC TEC Data Sharing / Barter Incentives Service"
authors = ["Jose Andrade <jose.r.andrade@inesctec.pt>", "Andre Garcia <andre.f.garcia@inesctec.pt>"]
license = "GNU Affero General Public License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.3"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
PyJWT = "^2.8.0"
pytz = "^2024.1"
psycopg2-binary = "^2.9.6"
pandas = "^2.2.1"
loguru = "^0.7.2"
gunicorn = "^23.0.0"
django-stronghold = "^0.4.0"
drf-yasg = "^1.21.7"
django-structlog = "^7.1.0"
django-cors-headers = "^4.3.1"
python-dotenv = "^1.0.1"
ruff = "^0.3.2"
setuptools = "^74.1.2"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.0.2"
pytest-django = "^4.8.0"
pytest-cov = "^4.1.0"
pytest-xdist = "^3.6.1"
pytest-html = "^4.1.1"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    "__pycache__",
    "docs",
    "*/migrations/",
    "*/tests/",
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
]
ignore = ["F401", "F405", "F403"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"